feat: builder update
one of builder's internal map variables was not initialized, which
made the controller fail to create tasks out of job silently while
only returning 500 error to flamectl.

Also, to test the builder code, the asyncfl_hier_mnist example was
updated accordingly. It was confirmed that the controller created the
configs for the example correctly.