fix: reset job builder state
job builder's state is not cleaned up after creating tasks out of a
job specification. In particular, variable 'channels' accumulates the
channel information across jobs, which is caught during the validation
check in the builder. To address the issue, a reset function is
defined and called every time tasks for a new job need to be created.